ProdleAI/requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.28.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
azure-core==1.32.0
azure-storage-blob==12.24.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.0.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
dataclasses-json==0.6.6
datasets==2.19.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
email_validator==2.1.1
environs==9.5.0
eval_type_backport==0.2.0
evaluate==0.4.2
exception==0.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.0
kubernetes==29.0.0
llama-cloud==0.0.17
llama-index==0.11.8
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.8
llama-index-embeddings-openai==0.2.4
llama-index-extractors-entity==0.1.2
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.3.0
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-chroma==0.2.0
llama-parse==0.5.3
llamaindex-py-client==0.1.19
lxml==5.2.2
mammoth==1.7.1
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
milvus_cli==1.0.1
minio==7.2.12
mkl==2021.4.0
ml-dtypes==0.4.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.44.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pillow==10.3.0
plum-dispatch==1.7.4
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymilvus==2.5.0
pymongo==4.7.2
PyMuPDF==1.24.7
pymupdf4llm==0.0.9
PyMuPDFb==1.24.6
pypandoc==1.13
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
redisvl==0.3.9
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seqeval==1.2.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
span-marker==1.5.0
Spire.Doc==12.7.1
Spire.Pdf==10.7.1
SQLAlchemy==2.0.30
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.6.0
together==1.2.7
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
